requirements/req.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/req.txt requirements/req.in
#
anyio==3.4.0
    # via starlette
asgiref==3.4.1
    # via uvicorn
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.9
    # via requests
click==8.0.3
    # via uvicorn
deprecated==1.2.13
    # via redis
fastapi==0.70.0
    # via -r requirements/req.in
h11==0.12.0
    # via uvicorn
httptools==0.3.0
    # via uvicorn
idna==3.3
    # via
    #   anyio
    #   requests
pydantic==1.8.2
    # via fastapi
python-dotenv==0.19.2
    # via uvicorn
python-multipart==0.0.5
    # via -r requirements/req.in
python-slugify==5.0.2
    # via -r requirements/req.in
pyyaml==6.0
    # via uvicorn
redis==4.0.2
    # via -r requirements/req.in
requests==2.26.0
    # via -r requirements/req.in
six==1.16.0
    # via python-multipart
sniffio==1.2.0
    # via anyio
starlette==0.16.0
    # via fastapi
text-unidecode==1.3
    # via python-slugify
typing-extensions==4.0.1
    # via pydantic
urllib3==1.26.7
    # via requests
uvicorn[standard]==0.16.0
    # via -r requirements/req.in
uvloop==0.16.0
    # via uvicorn
watchgod==0.7
    # via uvicorn
websockets==10.1
    # via uvicorn
wrapt==1.13.3
    # via deprecated
